test(ir): resolve the seam memory/config toggles via a shared support stub
The memory port made aimee_ir_serve.o reference ir_stage_memory +
gw_stage_memory_enabled + config_load + config_module_enabled. LTO --gc-sections
dropped the unreachable path locally, but CI's full unit-tests link surfaced
undefined references in unit-test-ir-legacy-parity (and would in
unit-test-aimee-ir-serve). Factor the four inert DISABLED stubs into
tests/support/ir_seam_memory_stub.c and link it into both minimal IR
build/parity suites (removing the inline copy from test_aimee_ir_serve.c). Full
'make unit-tests' + build-integrity green.
