+++ src/tests/test_aimee_ir_serve.c
@@ -10,35 +10,12 @@
 #include "agent_protocol.h"
 #include "aimee_ir.h"
 #include "aimee_ir_serve.h"
-#include "config.h"
 #include "cJSON.h"
 
-/* aimee_ir_serve.c registers the ported memory module on the transform seam, so it
- * now references the memory + config toggles. This suite exercises the build/
- * translation parity paths, NOT memory injection, so keep it a minimal link: stub
- * the module DISABLED (gw_stage_memory_enabled -> 0) with an inert transform, and a
- * config_load that reports "no config" so enablement falls to that env default. With
- * memory off the seam is a no-op and every byte-exact assertion below is unchanged. */
-int ir_stage_memory(aimee_request_t *ir, void *ud)
-{
-   (void)ir;
-   (void)ud;
-   return 0;
-}
-int gw_stage_memory_enabled(void)
-{
-   return 0;
-}
-int config_load(config_t *c)
-{
-   if (c)
-      memset(c, 0, sizeof *c);
-   return -1;
-}
-int config_module_enabled(int config_tristate, int env_default)
-{
-   return config_tristate >= 0 ? config_tristate : env_default;
-}
+/* The memory module + config toggles that aimee_ir_serve.c now references (memory is
+ * registered on the IR transform seam) are satisfied by tests/support/
+ * ir_seam_memory_stub.o -- stubbed DISABLED, so this build/translation-parity suite
+ * stays a minimal link and its byte-exact assertions are unchanged. */
 
 static const char *REQ =
     "{\"model\":\"claude-3-5-sonnet\",\"max_tokens\":100,"
